+++ test/templates/main.cpp
@@ -41,16 +41,16 @@ void test_decode(void) {
 }
 
 void test_signer_create_url_missing_nonce(void) {
-	LnurlSignerConfig config;
+	Lnurl::SignerConfig config;
 	config.apiKey.id = "5d4aeb462a";
 	config.apiKey.key = "ef9901bebc801518e7d862c2edaedd3acd86ec132fb3bd5ac0013c9a5ba478db";
 	config.apiKey.encoding = "hex";
 	config.callbackUrl = "https://localhost:3000/lnurl";
 	config.shorten = false;
-	LnurlSigner signer(config);
-	LnurlWithdrawParamsMSat params;
-	params.minWithdrawable = 50000;
-	params.maxWithdrawable = 60000;
+	Lnurl::Signer signer(config);
+	Lnurl::WithdrawParams params;
+	params.minWithdrawable = "50000";
+	params.maxWithdrawable = "60000";
 	params.defaultDescription = "";
 	try {
 		signer.create_url(params);
@@ -64,19 +64,18 @@ void test_signer_create_url_missing_nonce(void) {
 	TEST_FAIL_MESSAGE("Expected exception to be thrown");
 }
 
-void test_signer_create_url_withdraw_fiat_minWithdrawable_lt_zero(void) {
-	LnurlSignerConfig config;
+void test_signer_create_url_withdraw_minWithdrawable_lt_zero(void) {
+	Lnurl::SignerConfig config;
 	config.apiKey.id = "5d4aeb462a";
 	config.apiKey.key = "ef9901bebc801518e7d862c2edaedd3acd86ec132fb3bd5ac0013c9a5ba478db";
 	config.apiKey.encoding = "hex";
 	config.callbackUrl = "https://localhost:3000/lnurl";
-	config.fiatCurrency = "CZK";
 	config.shorten = false;
-	LnurlSigner signer(config);
+	Lnurl::Signer signer(config);
 	const std::string nonce = "test";
-	LnurlWithdrawParamsFiat params;
-	params.minWithdrawable = -1;
-	params.maxWithdrawable = 10000;
+	Lnurl::WithdrawParams params;
+	params.minWithdrawable = "-1";
+	params.maxWithdrawable = "10000";
 	params.defaultDescription = "";
 	try {
 		signer.create_url(params, nonce);
@@ -90,44 +89,18 @@ void test_signer_create_url_withdraw_fiat_minWithdrawable_lt_zero(void) {
 	TEST_FAIL_MESSAGE("Expected exception to be thrown");
 }
 
-void test_signer_create_url_withdraw_msat_maxWithdrawable_lt_minWithdrawable(void) {
-	LnurlSignerConfig config;
+void test_signer_create_url_withdraw_maxWithdrawable_lt_minWithdrawable(void) {
+	Lnurl::SignerConfig config;
 	config.apiKey.id = "5d4aeb462a";
 	config.apiKey.key = "ef9901bebc801518e7d862c2edaedd3acd86ec132fb3bd5ac0013c9a5ba478db";
 	config.apiKey.encoding = "hex";
 	config.callbackUrl = "https://localhost:3000/lnurl";
 	config.shorten = false;
-	LnurlSigner signer(config);
+	Lnurl::Signer signer(config);
 	const std::string nonce = "test";
-	LnurlWithdrawParamsMSat params;
-	params.minWithdrawable = 80000;
-	params.maxWithdrawable = 70000;
-	params.defaultDescription = "";
-	try {
-		signer.create_url(params, nonce);
-	} catch (const std::invalid_argument& e) {
-		TEST_ASSERT_EQUAL_STRING(
-			"\"maxWithdrawable\" must be greater than or equal to \"minWithdrawable\"",
-			e.what()
-		);
-		return;
-	}
-	TEST_FAIL_MESSAGE("Expected exception to be thrown");
-}
-
-void test_signer_create_url_withdraw_fiat_maxWithdrawable_lt_minWithdrawable(void) {
-	LnurlSignerConfig config;
-	config.apiKey.id = "5d4aeb462a";
-	config.apiKey.key = "ef9901bebc801518e7d862c2edaedd3acd86ec132fb3bd5ac0013c9a5ba478db";
-	config.apiKey.encoding = "hex";
-	config.callbackUrl = "https://localhost:3000/lnurl";
-	config.fiatCurrency = "CZK";
-	config.shorten = false;
-	LnurlSigner signer(config);
-	const std::string nonce = "test";
-	LnurlWithdrawParamsFiat params;
-	params.minWithdrawable = 75.00;
-	params.maxWithdrawable = 70.00;
+	Lnurl::WithdrawParams params;
+	params.minWithdrawable = "75000";
+	params.maxWithdrawable = "50000";
 	params.defaultDescription = "";
 	try {
 		signer.create_url(params, nonce);
@@ -142,16 +115,16 @@ void test_signer_create_url_withdraw_fiat_maxWithdrawable_lt_minWithdrawable(voi
 }
 
 void test_signer_create_url_custom_params_reserved(void) {
-	LnurlSignerConfig config;
+	Lnurl::SignerConfig config;
 	config.apiKey.id = "5d4aeb462a";
 	config.apiKey.key = "ef9901bebc801518e7d862c2edaedd3acd86ec132fb3bd5ac0013c9a5ba478db";
 	config.apiKey.encoding = "hex";
 	config.callbackUrl = "https://localhost:3000/lnurl";
-	LnurlSigner signer(config);
+	Lnurl::Signer signer(config);
 	const std::string nonce = "test";
-	LnurlWithdrawParamsMSat params;
-	params.minWithdrawable = 50000;
-	params.maxWithdrawable = 50000;
+	Lnurl::WithdrawParams params;
+	params.minWithdrawable = "50000";
+	params.maxWithdrawable = "50000";
 	params.defaultDescription = "";
 	params.custom["id"] = "cannot_use_reserved_keys";
 	try {
@@ -173,9 +146,8 @@ int main(void) {
 	RUN_TEST(test_encode);
 	RUN_TEST(test_decode);
 	RUN_TEST(test_signer_create_url_missing_nonce);
-	RUN_TEST(test_signer_create_url_withdraw_fiat_minWithdrawable_lt_zero);
-	RUN_TEST(test_signer_create_url_withdraw_msat_maxWithdrawable_lt_minWithdrawable);
-	RUN_TEST(test_signer_create_url_withdraw_fiat_maxWithdrawable_lt_minWithdrawable);
+	RUN_TEST(test_signer_create_url_withdraw_minWithdrawable_lt_zero);
+	RUN_TEST(test_signer_create_url_withdraw_maxWithdrawable_lt_minWithdrawable);
 	RUN_TEST(test_signer_create_url_custom_params_reserved);
 // {{SIGNER_RUN_TESTS}}
 	return UNITY_END();